requirements.txt
asgiref==3.5.0
Django==4.0.1
django-cors-headers @ file:///media/yagshy/Local%20Disk/share/APP/pypi/django-cors-headers-3.13.0.tar.gz
djangorestframework==3.13.1
pydivert==2.1.0
pytz==2021.3
scapy==2.5.0
sqlparse==0.4.2
